+++ src/main/kotlin/com/example/ajaxproject/repository/MentorRepository.kt
@@ -5,4 +5,4 @@ import com.example.ajaxproject.model.User
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
-interface MentorRepository : CrudRepository<Mentors, Int>
+interface MentorRepository : CrudRepository<Mentors, Long>
